separated details into own  component
